# nodeproject35/server.js
const express = require('express');
const app = express();
const cors = require('cors');
const mongoose = require('mongoose');
const multer = require('multer')
const Category = require('./models/category')
const User = require('./models/User')

const UserProfile = require('./models/UserProfile')
const Seller = require('./models/Seller')
const Brand = require('./models/brand')
const SellerProfile = require('./models/SellerProfile')
const Image = require('./models/Image')
const Product = require('./models/product')
const Cart = require('./models/Cart')
const UserOrders = require('./models/UserOrders')

// const MobileFeature = require('./models/MobileFeature')
// const LaptopFeature = require('./models/LaptopFeature')
// const HeadphonesFeature = require('./models/HeadphonesFeature')
const swaggerUi = require('swagger-ui-express')
const YAML = require('yamljs');
const product = require('./models/product');
const swaggerDocument = YAML.load('./swagger.yaml')

app.use(express.json())
app.use(cors());

app.use(
  "/api-docs",
  swaggerUi.serve,
  swaggerUi.setup(swaggerDocument)
);

mongoose.connect("mongodb+srv://madhesh:dobOct2001@cluster0.exwrl.mongodb.net/dreambasket?retryWrites=true&w=majority",{
    useNewUrlParser: true,
  })
mongoose.set('toJSON',{
  virtuals: true,
  transform: (doc,converted)=> {
    delete converted._id;
  }
})
const db = mongoose.connection;
db.on("error", console.error.bind(console, "connection error: "));
db.once("open", function () {
  console.log("Connected successfully");
});

const storage = multer.diskStorage({
  destination: (req, file, cb) => {
    cb(null, 'uploads')
  },
  filename : (req, file, cb) => {
    cb(null, file.originalname)
  }
})
const upload = multer({storage: storage}).single('testimage')


//User
app.get("/users",cors(),async (req,res)=>{
  const user = await User.find({})
  user_json= JSON.stringify(user)
  res.header('Content-Range','user 0-20/20')
  res.json(user)
})

app.post("/users",cors(),async (req,res)=>{

  const newUser = new User(req.body)
  const user = await newUser.save()
  res.json(user)
})

//Category
app.post("/categories",cors(),async (req,res)=>{
    const category = new Category(req.body)
    const cate = await category.save()
    res.json(cate)
})

app.get("/categories",cors(),async (req,res)=>{
  const cat = await Category.find({})
  
  res.header('Content-Range','category 0-20/20')
  res.json(cat)
})

app.get('/categories/:id',cors(), async (req,res)=>{
    const cat = await Category.findById({"_id": req.params.id})
    res.json(cat)
})

//Brand

app.post("/brands",cors(), async (req,res)=>{
  const bran = new Brand(req.body)
  const b = await bran.save()
  res.json(b)
})
app.get("/brands",cors(), async (req,res)=>{
  const bran = await Brand.find({})
  // const b = JSON.stringify(bran)
  res.header('Content-Range','brands 0-20/20')
  res.json(bran)
})

app.get('/brands/:id',cors(), async (req,res)=>{
  const brand_id = await Brand.findById({"_id": req.params.id})
  res.json(brand_id)
})
// Brands for Categories


app.get("/categories/:id/brands",cors(), async (req,res)=>{
  const bran = await Brand.find({Category: req.params.id})
  // const b = JSON.stringify(bran)
  res.json(bran)
})

//Profile
app.get("/user/:id/userprofile",cors(),async (req,res)=>
{
  const userProfile = await UserProfile.find({userId: req.params.id})
  // const userProf= JSON.stringify(userProfile)
  res.json(userProfile)
})





app.post("/users/:id/userprofile",cors(),async (req,res)=>{

  const userprofile= new UserProfile({...req.body,userId: req.params.id})
  const profile= await userprofile.save()
  res.json(profile)
})


//Seller
app.post("/sellers",cors(),async (req,res)=>{
  
  console.log(req.body);
  const seller_temp= new Seller(req.body)
  const seller_t = await seller_temp.save()
  res.json(seller_t)
})

app.get("/sellers",cors(),async (req,res)=>{
  
  const seller = await Seller.find({})
  res.header('Content-Range','sellers 0-20/20')
  res.json(seller)
})

app.get("/sellers/:id",cors(), async(req,res)=>{
  const seller = await Seller.findById({ "_id": req.params.id})
})


// Seller profile
app.post('/sellers/:id/sellerprofile',cors(), async(req,res)=>{
  const sellerprofile = new SellerProfile({...req.body,sellerId: req.params.id})
  const profile = await sellerprofile.save()
  res.json(profile)
})


app.get('/sellers/:id/sellerprofile',cors(), async(req,res)=>{

  const sellerprofile = await SellerProfile.find({sellerId: req.params.id})
  res.json(sellerprofile) 
})

// Features of a category
app.get('/categories/:id/features',cors(), (req,res)=>{

  var query = { cid: req.params.id };
  db.collection("features").find(query).toArray(function(err, result) {
    if (err) throw err;
    res.json(result);
    //db.close();
  });

  
})


// Product

app.get('/sellerproduct',cors(),async(req,res)=>{
  const products = await Product.find({})
  res.json(products)
})


app.post("/sellerproduct",cors(),async (req,res)=>{

  const newProduct= new Product(req.body)
  const product= await newProduct.save()
  res.json(product)
})

app.get('/categories/:cid/sellerproduct',cors(), async(req,res)=>{
  const products = await Product.find({"Category": req.params.cid})
  res.json(products)
})

app.get('/categories/:cid/brands/:brandname/sellerproduct',cors(),async(req,res)=>{
  try{
    const products= await Product.find({Category:req.params.cid, productbrand: req.params.brandname})
    res.json(products)
  }
  catch(err){
    console.log(err)
  }
  
})


//Image Upload

app.post('/product/upload',cors(),async(req, res)=>{
   upload(req, res, (err)=>{
    if(err){
      console.log(err)
    }
    else{
      // const newProduct = newProduct(req.body)


      const newImage = new Image({
        name : req.body.name,
        image: {
          data: req.file.filename,
          contentType: 'image/png'
        }
      })
      newImage.save().then(()=>{
        res.json(newImage)
      })
      // const product = Product.find({"_id":req.params.id})
      // product.images.push(img._id)
      // product.save(done)
      // res.json(img)
    }
  })
})

app.get('/product/upload',cors(),async(req,res)=>{
  // const product = Product.find({"_id":req.params.id}).populate(Image)
  const img = await Image.find({})
  res.json(img)
})


//Cart

app.get('/cart',cors(),async(req,res)=>{
  let cartItems = await Cart.find({})
  res.json(cartItems)
})

app.delete('/cart',cors(),async(req,res)=>{
  const cartItems = await Cart.deleteMany()
})

app.get('/users/:id/cart',cors(),async(req,res)=>{
  const cartItems = await Cart.find({userId: req.params.id})
  console.log(cartItems)
  res.json(cartItems)
})

app.post('/users/:id/cart',cors(),async(req,res)=>{
  const newCartItem = await new Cart({...req.body,userId: req.params.id,})
  const item = newCartItem.save()
  console.log(item)
  res.json(newCartItem)
})

app.get('/users/:id/cart/:cartId',cors(),async(req,res)=>{
  const cartItem = await Cart.find({_id: req.params.cartId})
  res.json(cartItem)
})

app.patch('/users/:id/cart/:cartId',cors(),async(req,res)=>{
  const cartItem = await Cart.findByIdAndUpdate({_id: req.params.cartId})
    cartItem.qty = req.body.qty
  const item = await cartItem.save()
  res.json(item)
})

app.delete('/users/:id/cart/:cartId',cors(),async(req,res)=>{
  // const cartItem = await Cart.findByIdAndRemove({_id: req.params.cartId})
  // res.json(cartItem)
  try{
    const cartItem = await Cart.deleteOne({_id: req.params.cartId})
    res.json(cartItem)
  }
  catch(error){
    console.log(error)
  }
})


//User Orders

app.get('/orders',cors(),async(req,res)=>{
  const orders = await UserOrders.find({})
  res.json(orders)

})


app.post('/users/:id/orders',cors(),async(req,res)=>{
  const newOrder = await new UserOrders(req.body)
  const order = await newOrder.save()
  console.log(order)
  res.json(order)
})

app.get('/users/:id/orders',cors(),async(req,res)=>{
  const orders = await UserOrders.find({userId:req.params.id})
  res.json(orders)

})

app.delete('/users/:id/orders/:orderId',cors(),async(req,res)=>{
  const orders = await UserOrders.deleteOne({_id:req.params.orderId})
  res.json(orders)

})

app.delete('/users/:id/orders',cors(),async(req,res)=>{
  const orders = await UserOrders.deleteMany({userId:req.params.id})
  res.json(orders)

})

app.get('/users/:id',cors(),async(req,res)=>{
  const user = User.find({_id:req.params.id})
  res.json(user)
})


// Seller Orders
app.get('/sellers/:id/orders',cors(),async(req,res)=>{
  const orders = await UserOrders.find({sellerId:req.params.id})
  res.json(orders)

})




app.listen(5000,()=>{
  console.log("listening......")
});