# nodeproject35/swagger.yaml
openapi : 3.0.0
info:
    title: 'BlogsPage API'
    version: '1.0.0'
    description: 'This is the documentation for the project'

servers: 
    - url : http://localhost:5000

tags:
  - name: Dream Basket
    description: API to manage e-commerce

paths:
  /brands:
    get:
      summary: List of all brands
      tags: [brand]
      responses:
        "200":
          description: The list of brands.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/brand'
  /categories/{categoryId}/brands:
    get:
      summary: Returns a brand by category  ID.
      tags: [brand]
      parameters:
        - name: categoryId
          in: path
          required: true
          description: Autogenerated category ID by MongoDB
          schema:
            type : string
      responses: 
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/brand'

  /categories:
    get:
      summary: List of all categ
      tags: [category]
      responses:
        "200":
          description: The list of categories.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/category'
  /categories/{categoryID}:
    get:
      summary: Returns a category by ID.
      tags: [category]
      parameters:
        - name: categoryID
          in: path
          required: true
          description: Autogenerated Student ID by MongoDB
          schema:
            type : string
      responses: 
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/category'
    post:
      summary: Creates a Category.
      tags: [category]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/category'
      responses:
        "200":
          description: successs
  /users:
    get:
      summary: List of all users
      tags: [user]
      responses:
        "200":
          description: The list of userss.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/users'
  /sellers:
    get:
      summary: List of all sellers
      tags: [seller]
      responses:
        "200":
          description: The list of sellers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/seller'
  
  
  
components:
  schemas:
    brand:
      type: object
      required:
        - Name
        - Category
        
      properties:
        id:
          type: integer
          description: The auto-generated id of the CATEGOEY.
        Name:
          type: string
          description: The name of the brand.
        Category:
          type: integer
          description: Brand category id
        
      example:
        Name: Boat
        Category:  id of Category
    sellerprofile:
      type: object
      required:
        - sellerId
        - firstName
        - lastName
        - PhNo
        - city
        - address
        
      properties:
        id:
          type: integer
          description: The auto-generated id of the CATEGORY.
        sellerId:
          type: integer
          description: The id of the seller .
        firstName:
          type: string
          description: firstName of seller
        lastName:
          type: string
          description: lastName of seller
        PhNo:
          type: integer
          description: phone number of seller
        city:
          type: string
          description: city of seller
        address:
          type: string 
          description: address of seller
      example:
        sellerId: Boat
        firstName:  Madhesh Eletronics
        lastName: Pvt Limited
        PhNo: 7798675432
        city: Palakol
        address: Opposite Axis Bank road
    
      
    category:
      type: object  
      properties:
        id:
          type: integer
          description: The auto-generated id of the category.
        Name:
          type: string
          description: The type of product in category.
        
      example:
        Name: mobilephone  

     
    seller:
      type: object
      required:
        - email
        - sellername
        - password
        - gst
        
      properties:
        id:
          type: integer
          description: The auto-generated id of the product.
        email:
          type: string
          description: The email of the seller .
        sellername:
          type: string
          description: The name of the seller 
        Brand:
          type: string
          description: The name of company the product belongs to 
        gst:
          type: string
          description: The C&SGSTs on the product
        
      example:
        Name: boAt Rockerz 450 
        Category: Headphones
        Brand: Boat
        gst: 12%
    users:
      type: object
      required:
        - email
        - username
        - password
        
        
      properties:
        id:
          type: integer
          description: The auto-generated id of the product.
        email:
          type: string
          description: The email of the User .
        username:
          type: string
          description: The name of the user 
        password:
          type: string
          description: The password that the user uses for login 
        
        
      example:
        email: ncshfv@gmail.com 
        username: Lilly
        password: Lilly890  
    cart:
      type: object
      required:
        - userId
        - cid
        - pid
        - sellerId
        - sellername
        - productname
        - productbrand
        - productprice
        - ram
        - storage
        - color
        - connectorType
        - productquantity
        - qty
        
        
      properties:
        id:
          type: integer
          description: The auto-generated id of the product.
        userId:
          type: string
          description: The ID of  the user .
        cid:
          type: string
          description: The ID of  the category .
        pid:
          type: string
          description: The ID of  the product .
        sellerId:
          type: string
          description: The ID of  the seller .
        sellername:
          type: string
          description: The name of the seller
        productname:
          type: string
          description: The name of the product.
        productbrand:
          type: string
          description: The brand of the product.
        productprice:
          type: number
          description: The product of the price.
        ram:
          type: string
          description: The ram of the product.
        storage:
          type: string
          description: The storage capacity of the product .
        color:
          type: string
          description: The color of the product.
        connectortype:
          type: string
          description: The connectortype of the product.
        productquantity:
          type: number
          description: The  available quantity of the product from the seller.
        qty:
          type: number
          description: The quantity of the product user wanna buy.
        
        
      # example:
      #  userId 
      #  cid
      #  pid
      #  sellerId
      #  sellername
      #  productname
      #  productbrand
      #  productprice
      #  ram
      #  storage
      #  color
      #  connectorType
      #  productquantity
      #  qty  
    


    